requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-docx==0.8.11
openpyxl==3.1.2
pdfplumber==0.10.3
pandas>=2.0.3
Werkzeug==2.3.7
numpy>=1.26.0
